wip: Add a more robust HTTP detection scheme
