+++ linkerd/proxy/http/Cargo.toml
@@ -11,6 +11,7 @@ This should probably be decomposed into smaller, decoupled crates.
 """
 
 [dependencies]
+async-trait = "0.1"
 bytes = "0.5"
 futures = { package = "futures", version = "0.3" }
 h2 = "0.2.6"
@@ -26,6 +27,7 @@ linkerd2-error = { path  = "../../error" }
 linkerd2-http-box = { path  = "../../http-box" }
 linkerd2-identity = { path  = "../../identity" }
 linkerd2-io = { path  = "../../io" }
+linkerd2-proxy-transport = { path  = "../transport" }
 linkerd2-stack = { path  = "../../stack" }
 linkerd2-timeout = { path  = "../../timeout" }
 rand = "0.7"
@@ -35,3 +37,6 @@ tracing = "0.1.19"
 tracing-futures = { version = "0.2", features = ["std-future"] }
 try-lock = "0.2"
 pin-project = "0.4"
+
+[dev-dependencies]
+tokio-test = "0.2"
