+++ linkerd/proxy/http/src/lib.rs
@@ -17,6 +17,7 @@ pub mod insert;
 pub mod normalize_uri;
 pub mod orig_proto;
 pub mod override_authority;
+pub mod server;
 pub mod strip_header;
 pub mod timeout;
 pub mod trace;
@@ -25,9 +26,9 @@ mod version;
 
 pub use self::{
     client_addr::{ClientAddr, SetClientAddr},
-    detect::DetectHttp,
     glue::{Body as Payload, HyperServerSvc},
     override_authority::CanOverrideAuthority,
+    server::ServeHttp,
     timeout::MakeTimeoutLayer,
     version::Version,
 };
