+++ linkerd/proxy/http/src/version.rs
@@ -21,8 +21,7 @@ impl std::convert::TryFrom<http::Version> for Version {
 }
 
 impl Version {
-    pub const DETECT_BUFFER_CAPACITY: usize = 8192;
-    const H2_PREFACE: &'static [u8] = b"PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n";
+    pub(crate) const H2_PREFACE: &'static [u8] = b"PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n";
 
     /// Tries to detect a known protocol in the peeked bytes.
     ///
