+++ linkerd/proxy/transport/Cargo.toml
@@ -15,6 +15,7 @@ mock-orig-dst  = []
 
 [dependencies]
 async-stream = "0.2.1"
+async-trait = "0.1"
 bytes = "0.5"
 futures = "0.3"
 indexmap = "1.0.0"
